pyproject.toml
[project]
name = "from-jira-to-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=2.6.11",
    "instructor>=1.7.4",
    "loguru>=0.7.3",
    "pre-commit>=4.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.10",
]
